+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## Release 1.1.1
 
 * Bugfix: discover for notTriggerableJobDefintions
+* Make size of jobresult history configurable through edison.jobtrigger.jobresults.max
 
 ## Release 1.1.0